[VFX][Feature] Six-way smoke lighting
This PR creates a new _Material Type_ for the VFX Lit Output of HDRP : SixWaySmokeLit. This feature gives the user the ability to achieve visually realistic smoke effects that react acurately to the lighting of the scene.

Two textures, baked from an external DDC (e.g. Embergen, Houdini, Blender), store the smoke response to lights coming from 6 directions (Right, Top, Back, Left, Bottom, Front). At runtime, these values are interpolated depending on the relative position of the light source.
This lighting model is integrated to the light loop of HDRP, so the smoke can be affected by all the different light types : 

- Point/Spot Light
- Directional Light
- Area Lights
- Ambient Lighting (LPPV, APV)

 
Additionally, there is an option to store the emission intensity of the smoke in the fourth channel of one of the textures. This intensity value can be remapped with a gradient to render flames. 
It is also possible to use a separate Emissive map that store RGB values.

Old, but not out-dated presentation : https://docs.google.com/presentation/d/1Q-xgFFXLIu54AwzSXwhy57woOy1ghNIPv-anFzEaCTA/edit#slide=id.ga990b2f242_0_35

JIRA Link : https://jira.unity3d.com/browse/VFXG-225

![SmokeLitGif](https://media.github.cds.internal.unity3d.com/user/2768/files/5f5c69b2-8f7f-4460-9d81-bb831fcfbb64)