URP performance test for 22LTS
Jira ticket: GFXRPF-209 - https://jira.unity3d.com/browse/GFXRPF-209
Create performance test project for LTS 2022. Port current Universal_PerformanceTests to execute the same performance metrics in 2022.
The tests creates an artifact with the output images too.
This is not a graphics tests, so the results between iterations will not be checked