[Port] [2022.3] Rendering Debugger - Keep the selected panel when entering playmode and the panel count changes.
There are debug panels that only appear when the editor is in playmode. Such as Display Stats.

Display stats, always appears on top of the Rendering Debugger panels list. This makes the indexes shift. The editor debug window is storing the id of the selected panel, and not the name. Names do not change between entering/exiting playmode. So is safer to keep the display name, And retrieve the current panel index from the DebugManager when requested.