[Backport][2022.3][SRPF] Use FNV-1a 32 for caching and resource hashcode in Render Graph
The main issue goes to the hash method generation that was used for identifying Render Graph resources. We used the most performant way but it was too many collisions. After a bit of research the decision is to use FNV-1a implementation as it's the same performance but much more reliable. There's still a possibility for collision but a probability should be much smaller.